# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: API for managing bank cards
  version: 1.0.0
  contact:
    name: Timofey Zaytsev
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /user/cards:
    get:
      tags:
        - User Cards
      summary: Get user cards
      description: Retrieves paginated list of cards for the authenticated user
      operationId: getUserCards
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Number of items per page
        - name: cardType
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardType'
          description: Filter by card type
        - name: currency
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Currency'
          description: Filter by currency
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
          description: Filter by card status
      responses:
        '200':
          description: Cards found
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/FindCardResponse'
                  pageable:
                    type: object
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  last:
                    type: boolean
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /user/cards/{cardId}/applications:
    post:
      tags:
        - User Cards
        - Applications
      summary: Create application
      description: Creates a new application for a specific card (e.g., card deletion, activation, blocking)
      operationId: createApplication
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Card ID
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application form created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        '400':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Forbidden - User does not have permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: Application form already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /user/cards/{cardId}/balance:
    put:
      tags:
        - User Cards
      summary: Top up card balance
      description: Adds funds to the specified card balance
      operationId: topUpCardBalance
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Card ID to top up
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopUpBalanceRequest'
      responses:
        '204':
          description: Balance topped up successfully
        '400':
          description: Validation exception (invalid amount or request body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Card is blocked or deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found or does not belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    get:
      tags:
        - User Cards
      summary: Get card balance
      description: Retrieves current balance for a specific card
      operationId: getCardBalance
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Card ID
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '403':
          description: Card is blocked or deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found or does not belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /user/cards/transfer:
    post:
      tags:
        - User Cards
      summary: Transfer between cards
      description: Transfers funds from one card to another
      operationId: transferBetweenCards
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Validation exception or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '403':
          description: Card is blocked, deleted, or different currency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
  schemas:
...
    FindCardResponse:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        maskedNumber:
          type: string
        cardType:
          $ref: '#/components/schemas/CardType'
        balance:
          type: number
          format: decimal
          description: Current card balance
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          $ref: '#/components/schemas/CardStatus'
        createdAt:
          type: string
          format: date-time
          description: Card creation timestamp
        expiryDate:
          type: string
          format: date
          description: Card expiration date (YYYY-MM-DD)
...
    TransferRequest:
      type: object
      required:
        - sourceCardId
        - toCardId
        - amount
      properties:
        sourceCardId:
          type: string
          format: uuid
          description: Source card ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        targetCardId:
          type: string
          format: uuid
          description: Destination card ID
          example: "987fcdeb-51a2-43d7-9b15-876543210000"
        amount:
          type: number
          format: decimal
          description: Transfer amount
          minimum: 0.01
          example: 100.50
        description:
          type: string
          example: "Some reason to transfer amount"
    BalanceResponse:
      type: object
      properties:
        maskedNumber:
          type: string
          description: Masked card number
          example: "**** **** **** 1234"
        balance:
          type: number
          format: decimal
          description: Current card balance
          example: 1500.50
        status:
          $ref: '#/components/schemas/CardStatus'
        currency:
          $ref: '#/components/schemas/Currency'
    TransferResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          description: Transaction ID
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
    TopUpBalanceRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: decimal
          description: Amount to top up
          minimum: 0.01
          example: 100.50
        currency:
          $ref: '#/components/schemas/Currency'
...
    CreateApplicationRequest:
      type: object
      required:
        - title
        - description
        - type
      properties:
        title:
          type: string
          description: Application title
          minLength: 5
          maxLength: 200
          example: "Card Deletion Request"
        description:
          type: string
          description: Detailed description of the application
          minLength: 5
          maxLength: 1000
          example: "I would like to request deletion of my credit card ending with 1234"
        type:
          $ref: '#/components/schemas/ApplicationType'
    CreateApplicationResponse:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
          description: Created application ID
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    ApplicationType:
      type: string
      enum:
        - DELETE_CARD
        - ACTIVATE_CARD
        - BLOCK_CARD
        - ANOTHER_REQUEST
      description: Type of application
      example: "DELETE_CARD"
...
    CardType:
      type: string
      enum:
        - MASTERCARD
        - VISA
        - MIR
      description: Type of the card
    Currency:
      type: string
      enum:
        - USD
        - EUR
        - RUB
      description: Currency code
    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - DELETED
        - DEADLINE_HAS_EXPIRED
      description: Current status of the card
...
    ErrorResponseDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        uri:
          type: string
          example: "/api/v1/admin/cards"